bug fix for previous change

+++ loss.py
@@ -72,7 +72,7 @@ def loss(predictions, labels, loss_type, int_lbl):
       labels_shape=tf.shape(labels)
       labels = tf.cast(labels,tf.float32)
       loss=tf.nn.l2_loss(tf.subtract(predictions ,labels))
-      loss=tf.truediv(loss,labels_shape[0]*labels_shape[1],name="l2_mean")
+      loss=tf.truediv(loss,tf.cast(labels_shape[0]*labels_shape[1],tf.float32),name="l2_mean")
     tf.add_to_collection('loss', loss)
   return loss
